requirements.txt
asn1crypto==0.21.1
attrs==16.3.0
Automat==0.5.0
cffi==1.9.1
colorama==0.3.7
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
decorator==4.0.11
flake8==3.3.0
idna==2.5
incremental==16.10.1
ipython==5.3.0
ipython-genutils==0.2.0
lxml==3.7.3
mccabe==0.6.1
packaging==16.8
parsel==1.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.13
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycodestyle==2.3.1
pycparser==2.17
pycrypto==2.6.1
PyDispatcher==2.0.5
pyflakes==1.5.0
Pygments==2.2.0
pyOpenSSL==16.2.0
pyparsing==2.2.0
queuelib==1.4.2
requests==2.13.0
Scrapy==1.3.3
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
Twisted==17.1.0
w3lib==1.17.0
wcwidth==0.1.7
win-unicode-console==0.5
zope.interface==4.3.3
